Reorganise a bit the testing infrastructure.
